add validation when using removeTeacherFromCourse
- teacher can not be removed from his last assigned course

+++ university/src/main/java/ua/foxminded/university/dao/service/TeacherService.java
@@ -61,6 +61,9 @@ public class TeacherService {
 	}
 
 	public int removeTeacherFromCourse(int teacherId, String courseName) {
+		if (teacherRepository.findTeachersRelatedToCourse(courseName).size() == 1) {
+			throw new IllegalStateException("Teacher must be assigned at least at one Course!");
+		}
 		return teacherRepository.removeTeacherFromCourse(teacherId, courseName);
 	}
